# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XHS AI Tool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      parameters:
        - name: region
          in: query
          schema:
            type: string
            enum: [SG, HK, MY, ALL]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, banned, all]
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Account deleted
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List assets
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [image, video, all]
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Register an uploaded asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Asset deleted
  /ai/rewrite:
    post:
      operationId: aiRewriteContent
      tags: [ai]
      summary: AI rewrite content based on reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRewriteBody"
      responses:
        "200":
          description: Rewritten content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRewriteResponse"
  /ai/check-sensitivity:
    post:
      operationId: aiCheckSensitivity
      tags: [ai]
      summary: Check content for sensitive words
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiCheckSensitivityBody"
      responses:
        "200":
          description: Sensitivity check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCheckSensitivityResponse"
  /ai/generate-title:
    post:
      operationId: aiGenerateTitle
      tags: [ai]
      summary: AI generate catchy titles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGenerateTitleBody"
      responses:
        "200":
          description: Generated titles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiGenerateTitleResponse"
  /ai/generate-hashtags:
    post:
      operationId: aiGenerateHashtags
      tags: [ai]
      summary: AI generate relevant hashtags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGenerateHashtagsBody"
      responses:
        "200":
          description: Generated hashtags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiGenerateHashtagsResponse"
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        region:
          type: string
          enum: [SG, HK, MY]
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, banned]
        notes:
          type: ["string", "null"]
        contentCount:
          type: integer
        lastActiveAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nickname
        - region
        - status
        - contentCount
        - createdAt
        - updatedAt
    CreateAccountBody:
      type: object
      properties:
        nickname:
          type: string
        region:
          type: string
          enum: [SG, HK, MY]
        avatarUrl:
          type: string
        notes:
          type: string
      required:
        - nickname
        - region
    UpdateAccountBody:
      type: object
      properties:
        nickname:
          type: string
        region:
          type: string
          enum: [SG, HK, MY]
        avatarUrl:
          type: string
        status:
          type: string
          enum: [active, inactive, banned]
        notes:
          type: string
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [image, video]
        filename:
          type: string
        objectPath:
          type: string
        size:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - filename
        - objectPath
        - size
        - tags
        - createdAt
    CreateAssetBody:
      type: object
      properties:
        accountId:
          type: integer
        type:
          type: string
          enum: [image, video]
        filename:
          type: string
        objectPath:
          type: string
        size:
          type: integer
        tags:
          type: array
          items:
            type: string
      required:
        - type
        - filename
        - objectPath
        - size
    AiRewriteBody:
      type: object
      properties:
        originalContent:
          type: string
        style:
          type: string
          enum: [casual, professional, creative, viral]
        region:
          type: string
          enum: [SG, HK, MY]
        additionalInstructions:
          type: string
      required:
        - originalContent
    AiRewriteResponse:
      type: object
      properties:
        rewrittenTitle:
          type: string
        rewrittenBody:
          type: string
        suggestedTags:
          type: array
          items:
            type: string
      required:
        - rewrittenTitle
        - rewrittenBody
        - suggestedTags
    AiCheckSensitivityBody:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
      required:
        - title
        - body
    AiCheckSensitivityResponse:
      type: object
      properties:
        score:
          type: number
        issues:
          type: array
          items:
            $ref: "#/components/schemas/SensitivityIssue"
        suggestion:
          type: string
      required:
        - score
        - issues
        - suggestion
    SensitivityIssue:
      type: object
      properties:
        word:
          type: string
        reason:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        suggestion:
          type: string
      required:
        - word
        - reason
        - severity
        - suggestion
    AiGenerateTitleBody:
      type: object
      properties:
        body:
          type: string
        style:
          type: string
          enum: [casual, professional, creative, viral]
        count:
          type: integer
          default: 5
      required:
        - body
    AiGenerateTitleResponse:
      type: object
      properties:
        titles:
          type: array
          items:
            type: string
      required:
        - titles
    AiGenerateHashtagsBody:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        count:
          type: integer
          default: 10
      required:
        - title
        - body
    AiGenerateHashtagsResponse:
      type: object
      properties:
        hashtags:
          type: array
          items:
            type: string
      required:
        - hashtags
...