# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XHS AI Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetType } from "./assetType";

export interface Asset {
  id: number;
  /** @nullable */
  accountId?: number | null;
  type: AssetType;
  filename: string;
  objectPath: string;
  size: number;
  tags: string[];
  createdAt: Date;
}