# lib/api-zod/src/generated/types/sensitivityIssue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XHS AI Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { SensitivityIssueSeverity } from "./sensitivityIssueSeverity";

export interface SensitivityIssue {
  word: string;
  reason: string;
  severity: SensitivityIssueSeverity;
  suggestion: string;
}